+++ bgw-core/src/main/kotlin/tools/aqua/bgw/builder/AnimationBuilder.kt
@@ -75,7 +75,7 @@ internal class AnimationBuilder {
 					onFinished(scene, animation)
 				}
 			}
-
+		
 		/**
 		 * Builds [MovementAnimation].
 		 */
@@ -88,12 +88,12 @@ internal class AnimationBuilder {
 			//Move node to initial position
 			node.layoutX = anim.fromX
 			node.layoutY = anim.fromY
-
+			
 			//set transition as relative movement
 			return TranslateTransition(Duration.millis(anim.duration.toDouble()), node).apply {
 				byX = anim.toX - anim.fromX
 				byY = anim.toY - anim.fromY
-
+				
 				//set on finished
 				onFinished = EventHandler {
 					node.layoutX = anim.toX
@@ -104,7 +104,7 @@ internal class AnimationBuilder {
 				}
 			}
 		}
-
+		
 		/**
 		 * Builds [RotationAnimation].
 		 */
@@ -120,7 +120,7 @@ internal class AnimationBuilder {
 			//set transition as relative movement
 			return RotateTransition(Duration.millis(anim.duration.toDouble()), node).apply {
 				byAngle = anim.toAngle - anim.fromAngle
-
+				
 				//set on finished
 				onFinished = EventHandler {
 					node.rotate = anim.toAngle
@@ -130,7 +130,7 @@ internal class AnimationBuilder {
 				}
 			}
 		}
-
+		
 		/**
 		 * Builds [FlipAnimation].
 		 */
@@ -143,18 +143,18 @@ internal class AnimationBuilder {
 			//Set initial scale
 			node.scaleX = anim.fromScaleX
 			node.scaleY = anim.fromScaleY
-
+			
 			return ScaleTransition(Duration.millis(anim.duration.toDouble()), node).apply {
 				toX = anim.toScaleX
 				toY = anim.toScaleY
-
+				
 				//set on finished
 				onFinished = EventHandler {
 					onFinished(scene, anim)
 				}
 			}
 		}
-
+		
 		/**
 		 * Builds [FlipAnimation].
 		 */
@@ -166,17 +166,17 @@ internal class AnimationBuilder {
 
 			//Set initial opacity
 			node.opacity = anim.fromOpacity
-
+			
 			return FadeTransition(Duration.millis(anim.duration.toDouble()), node).apply {
 				fromValue = anim.fromOpacity
 				toValue = anim.toOpacity
-
+				
 				onFinished = EventHandler {
 					onFinished(scene, anim)
 				}
 			}
 		}
-
+		
 		/**
 		 * Builds [FlipAnimation].
 		 */
@@ -187,7 +187,7 @@ internal class AnimationBuilder {
 			val node = mapNode(scene, anim.componentView)
 			val fromVisual = VisualBuilder.buildVisual(anim.fromVisual)
 			val toVisual = VisualBuilder.buildVisual(anim.toVisual).apply { scaleX = 0.0 }
-
+			
 			val animation1 = ScaleTransition(Duration.millis(anim.duration / 2.0), fromVisual).apply {
 				fromX = 1.0
 				toX = 0.0
@@ -196,22 +196,22 @@ internal class AnimationBuilder {
 				fromX = 0.0
 				toX = 1.0
 			}
-
-
+			
+			
 			node.children[0] = fromVisual
 			animation1.setOnFinished {
 				node.children[0] = toVisual
 				animation2.play()
 			}
-
+			
 			//set on finished
 			animation2.setOnFinished {
 				onFinished(scene, anim)
 			}
-
+			
 			return animation1
 		}
-
+		
 		/**
 		 * Builds [DelayAnimation].
 		 */
@@ -235,7 +235,7 @@ internal class AnimationBuilder {
 			anim: DiceAnimation<*>
 		): Transition {
 			val seq = SequentialTransition()
-
+			
 			repeat(anim.speed) {
 				seq.children += PauseTransition(Duration.millis(anim.duration / anim.speed.toDouble())).apply {
 					onFinished = EventHandler {
@@ -243,15 +243,15 @@ internal class AnimationBuilder {
 					}
 				}
 			}
-
+			
 			seq.setOnFinished {
 				anim.componentView.currentSide = anim.toSide
 				onFinished(scene, anim)
 			}
-
+			
 			return seq
 		}
-
+		
 		/**
 		 * Builds [RandomizeAnimation].
 		 */
@@ -260,7 +260,7 @@ internal class AnimationBuilder {
 			anim: RandomizeAnimation<*>
 		): Transition {
 			val seq = SequentialTransition()
-
+			
 			repeat(anim.speed) {
 				seq.children += PauseTransition(Duration.millis(anim.duration / anim.speed.toDouble())).apply {
 					setOnFinished {
@@ -268,21 +268,21 @@ internal class AnimationBuilder {
 					}
 				}
 			}
-
+			
 			seq.setOnFinished {
 				(anim.componentView as CardView).backVisual = anim.toVisual
 				onFinished(scene, anim)
 			}
-
+			
 			return seq
 		}
-
+		
 		/**
 		 * Removes [anim] from animations list and invokes [Animation.onFinished].
 		 */
 		private fun onFinished(scene: Scene<out ComponentView>, anim: Animation) {
 			scene.animations.remove(anim)
-
+			
 			Platform.runLater {
 				anim.onFinished?.invoke(AnimationFinishedEvent())
 			}